+++ Changelog.md
@@ -4,8 +4,24 @@ Expect active development and potentially significant breaking changes in the `0
 
 ### vNext
 
+### 1.0.0-rc.4
+- Included cached data to the child component along with the error. [PR #548]
+
+### 1.0.0-rc.3
+- Fix bug where `options` was mutated causing variables to not update appropriately. [PR #537](https://github.com/apollographql/react-apollo/pull/537)
+- Make sure that all queries resolve or reject if an error was thrown when server side rendering. [PR #488](https://github.com/apollographql/react-apollo/pull/488)
+- ApolloProvider now changes its client and store when those props change. [PR #479](https://github.com/apollographql/react-apollo/pull/479)
+
+### 1.0.0-rc.1
+- Update dependency to Apollo Client 1.0.0-rc.1 [PR #520](https://github.com/apollographql/react-apollo/pull/520)
+
+### 0.13.3
+- Make sure that the cached rendered element has the correct type before returning it. [PR #505](https://github.com/apollographql/react-apollo/pull/505)
+- Move constructor initializing of props to componentWillMount. [PR #506](https://github.com/apollographql/react-apollo/pull/506) ([Issue #509](https://github.com/apollographql/react-apollo/issues/509)).
+
 ### 0.13.2
-- Included cached data to the child component along with the error
+- Address deprecation warnings coming from `graphql-tag` [graphql-tag#54](https://github.com/apollographql/graphql-tag/issues/54)
+- Make sure ApolloClient and gql are exported from browser bundle [PR #501](https://github.com/apollographql/react-apollo/pull/501)
 
 ### 0.13.1
 - Add apollo-client ^0.10.0 to dependency range